mv astrometry iraf_source_detection to tools.py
